refactor: renamed dir_packager to directory_packager
BREAKING CHANGE:
